+++ R/unsmooth_gmap.R
@@ -1,9 +1,7 @@
-#' Smooth genetic map
+#' Unsmooth genetic map
 #'
-#' Smooth a genetic map by mixing it with a bit of constant
-#' recombination (using a separate recombination rate for each
-#' chromosome), to eliminate intervals that have exactly 0
-#' recombination.
+#' Performs the reverse operation of [smooth_gmap()], in case one wants to go back
+#' to the original genetic map.
 #'
 #' @param gmap Genetic map, as a list of numeric vectors; each vector gives marker
 #' positions for a single chromosome.
@@ -11,25 +9,26 @@
 #' positions for a single chromosome, with the same chromosomes and markers as `gmap`.
 #' @param alpha Proportion of mixture to take from constant recombination.
 #'
-#' @return A genetic map like the input `gmap`, but smoothed by mixing
-#' it with a proportion `alpha` of constant recombination on each
-#' chromosome.
+#' @return A genetic map like the input `gmap`, but with the reverse
+#' operation of [smooth_gmap()] applied, provided that exactly the
+#' same physical map and `alpha` are used.
 #'
 #' @details An interval of genetic length \eqn{d_g}{dg} and physical
 #'     length \eqn{d_p}{dp} is changed to have length
-#'     \eqn{(1-\alpha)d_g + \alpha d_p r}{(1-alpha)dg + (alpha*dp*r)}
+#'     \eqn{(d_g - \alpha d_p r)/(1-\alpha)}{(dg + alpha*dp*r)/(1-alpha)}
 #'     where \eqn{r = L_g / L_p}{r = Lg/Lp} is the chromosome-specific
 #'     recombination rate.
 #'
 #' @examples
 #' iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2"))
 #' gmap_adj <- smooth_gmap(iron$gmap, iron$pmap)
+#' gmap_back <- unsmooth_gmap(gmap_adj, iron$pmap)
 #'
 #' @export
 #' @keywords utilities
-#' @seealso [unsmooth_gmap()]
+#' @seealso [smooth_gmap()]
 
-smooth_gmap <-
+unsmooth_gmap <-
     function(gmap, pmap, alpha=0.02)
     {
         stopifnot(is.list(gmap), is.list(pmap))
@@ -42,7 +41,7 @@ smooth_gmap <-
             gd <- diff(g)
             pd <- diff(p)
             rr <- diff(range(g)) / diff(range(p))
-            gd <- gd*(1-alpha) + alpha*pd*rr
+            gd <- (gd- alpha*pd*rr)/(1-alpha)
             gmap[[chr]] <- setNames(cumsum(c(g[1], gd)), names(g))
         }
         gmap